Use the path to control loading tests
Currently, the app is reloaded when one of the test case links is clicked. This is currently a benefit, as there is no mechanism for removing the old test case from the map.

In the future, the click should be captured, the old test case discarded and the new one loaded, with the URL being changed with `window.history.pushState`, so that the page doesn't have to reload, but the URL still reflects the state.