Merge Lane and Buffer, removing the E indirection
The more I think about it, the difference between a "buffer" and a "lane" is a spectrum, with no cutoff that makes sense in all situations, and a lot of overlapping attributes. I am thinking that a homogeneous representation, with optional sub-structs for variant-specific data makes sense.

`Lane` should be renamed to reflect that it represent all road areas, not just lanes.