Simplify IntersectionControl, use IntersectionKind::MapEdge instead of Border
Border has no place in IntersectionControl now that MapEdge covers it. Given the scope of IntersectionControl, as a summary of an entire intersection, I think the vagueness of `Signalled` and `Signed` is appropriate. The kind of sign at each arm of the intersection will need to live on the `Road` itself.

This was a pretty mechanical find-replace. The test diffs show the rename of StopSign -> Signed, etc. and filling in a lot of Unknown intersection types in `road_network.dot`.