Package the native lib and load it from the .jar
The Java class has been put in the `org.osm2streets` package.
The build script packages the `.class` and `.so` files into a `.jar`.
The native lib is unpacked and loaded directly, instead of searching `java.library.path`.

This iteration only supports linux because it packages `target/debug/osm2streets_java.so` and loads it without checking the current platform. The next iteration should package different target builds and load the one appropriate for the running system.