Replace `country-geocoder` with `country-boundaries`
Instead of relying on [`country-geocoder`](https://github.com/a-b-street/country-geocoder)
to find a country code, we now use
[`country-boundaries`](https://github.com/westnordost/country-boundaries-rust/).
This makes the lookup faster, but more importantly it also returns
correct results for regions and countries which are not
`admin_level=0`, such as Hong Kong or Macau.

`country-boundaries` does not include data on whether a country
drives on the left or right, however thankfully `muv-osm` does, so
we use its data instead (which is generated from `wikidata.org`).